devops/chat_api_llm/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.24.1
python-multipart==0.0.6
langchain==0.1.9
google-generativeai>=0.4.1,<0.5.0
langchain-google-genai==0.0.11 